added constructors and the ability to change filepath
added two constructors. one finds the file in the relative director
the second finds it in a given path
refractored the created of the inputstream for better readability
added function to get the a block of data to be sent as a Dgram
parseFilePath turns a regular path string to one that the InputStream
object can understand by getting rid of the "\" issue with strings
the parse function has not been implemented yet.