Fixed lost Data on RRQ and WRQ one after another
startNewTransfer is used to see if at the end of a read the sim should listen to the client or not
cause the last thing on a read is the ack going to the server. otherwise the sim would just wait on server when the client tried to send a new file

endOfWRQ is used to see if the posrt should be set back to 69. before after one transfer the sim would still send the next file
to the previous server port. but all new transfers need to go to port 69

The fix involves two booleans and might not be the best method to solve it, if you find another way, feel free to use that.
But i have tested lost data with this and it works perfectly on read and write one after another.