Issue #28 Fixed
Issue with this was that the sim was basically listening to the wrong side. changing the side came with changing the ErrorSimState aswell and resetting it once the transfer is over
The sim also has some byte to int conversion issues, #29 refers to this. the //TODO comments correspond to the lines i am talking about

errorSimMode.java needed a set state method so i can change the state back and forth

The Client was not dealing with duplicate data properly
It was recognizing it, creates a new ack and nothing after that, doesnt send it anywhere or anything
So i changed the block number checks to just see if the block number is the next block number, if not, discard and send an ack for it.
i left what i dont use in the comments so you know what i did